API bug fixes and other enhancements

+++ src/main/java/tracker/model/AircraftFlightData.java
@@ -2,6 +2,6 @@ package tracker.model;
 
 public record AircraftFlightData(int alt, String cou, int gnd, String icao,
                                  float lat, float lon, String reg,
-                                 int spd, int talt, short trak, String type,
+                                 float spd, int talt, float trak, String type,
                                  int vsi) {
 }
